+++ src/main/kotlin/io/snyk/plugin/Utils.kt
@@ -14,6 +14,7 @@ import com.intellij.openapi.options.ShowSettingsUtil
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectRootManager
+import com.intellij.openapi.util.io.toNioPathOrNull
 import com.intellij.openapi.util.registry.Registry
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.wm.ToolWindow
@@ -59,6 +60,7 @@ import java.net.URI
 import java.nio.file.Path
 import java.security.KeyStore
 import java.util.Objects.nonNull
+import java.util.SortedSet
 import java.util.concurrent.TimeUnit
 import javax.net.ssl.SSLContext
 import javax.net.ssl.TrustManager
@@ -403,3 +405,12 @@ fun getArch(): String {
 fun VirtualFile.toPsiFile(project: Project): PsiFile? {
     return PsiManager.getInstance(project).findFile(this)
 }
+
+fun Project.getContentRootPaths(): SortedSet<Path> {
+    return getContentRootVirtualFiles()
+        .mapNotNull { it.path.toNioPathOrNull() }
+        .toSortedSet()
+}
+
+fun Project.getContentRootVirtualFiles() = ProjectRootManager.getInstance(this).contentRoots
+    .filter { it.exists() && it.isDirectory }.toSet()